Solved it, Byte just doesn't like having 8 bits.

+++ src/Main.java
@@ -342,7 +342,7 @@ class DES {
 
     }
 
-    public Byte encryptAndDecrypt(Byte input, boolean isEncryptionMode){
+    public int encryptAndDecrypt(int input, boolean isEncryptionMode){
 
         if(verboseTest && isEncryptionMode){System.out.println("Encryption has begun.");}
         if(verboseTest && !isEncryptionMode){System.out.println("Decryption has begun.");}
@@ -399,7 +399,7 @@ class DES {
         }
 
         System.out.println(answerString);
-        Byte byteAnswer = Byte.valueOf(answerString, 2);
+        int byteAnswer = Integer.valueOf(answerString, 2);
 
         if(verboseTest && isEncryptionMode){
             System.out.println("Encryption was used on input " + input + "\n" +
@@ -420,7 +420,7 @@ public class Main {
     public static void main(String[] args) {
 
         DES blockCypher = new DES("0101000100");
-        Byte cyphertext = blockCypher.encryptAndDecrypt(Byte.valueOf("101011",2), true);
+        int cyphertext = blockCypher.encryptAndDecrypt(Byte.valueOf("101011",2), true);
         blockCypher.encryptAndDecrypt(cyphertext, false);
 
         if(false) {
